python_and_r/NLP/requirements.txt
appdirs==1.4.4
argcomplete==1.10.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bayoo-docx==0.2.17
beautifulsoup4==4.9.3
bleach==3.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
docx2txt==0.8
EbookLib==0.17.1
entrypoints==0.3
et-xmlfile==1.0.1
extract-msg==0.23.1
filelock==3.0.12
html5lib==1.1
idna==2.10
IMAPClient==2.1.0
importlib-metadata==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
language-tool-python==2.5.2
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.20.1
olefile==0.46
openpyxl==3.0.6
packaging==20.7
pandas==1.2.2
pandocfilters==1.4.3
parso==0.7.1
pdfminer.six==20181108
pickleshare==0.7.5
Pillow==8.1.0
prettytable==2.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.7.2
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-docx==0.8.10
python-pptx==0.6.18
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.1
scipy==1.6.0
Send2Trash==1.5.0
six==1.12.0
sortedcontainers==2.3.0
soupsieve==2.2
SpeechRecognition==3.8.1
terminado==0.9.1
testpath==0.4.4
textract==1.6.3
tika==1.24
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
tzlocal==1.5.1
urllib3==1.26.3
virtualenv==20.2.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
XlsxWriter==1.3.7
zipp==3.4.0
